+++ glue/viewers/image/state.py
@@ -69,11 +69,11 @@ class ImageViewerState(MatplotlibDataViewerState):
 
         self.xw_att_helper = ComponentIDComboHelper(self, 'x_att_world',
                                                     numeric=False, categorical=False,
-                                                    visible=False, world_coord=True)
+                                                    world_coord=True)
 
         self.yw_att_helper = ComponentIDComboHelper(self, 'y_att_world',
                                                     numeric=False, categorical=False,
-                                                    visible=False, world_coord=True)
+                                                    world_coord=True)
 
         self.add_callback('reference_data', self._reference_data_changed, priority=1000)
         self.add_callback('layers', self._layers_changed, priority=1000)